servers: Wait for source topics to be created
Previously, a Kafka Streams application could exit in
case a source topic was not present when required for processing.
This was causing the kafka cloud server application to
often crash-loop for a while.

This commit ensures that the demo kafka server application will
wait till a needed source topic is created before processing its
messages.